+++ test/byContract/IexecPocoBoost/IexecPocoBoostDelegate.test.ts
@@ -51,7 +51,6 @@ const dealTagTee = '0x0000000000000000000000000000000000000000000000000000000000
 const standardDealTag = '0x0000000000000000000000000000000000000000000000000000000000000000';
 const taskIndex = 0;
 const volume = taskIndex + 1;
-const startTime = 9876543210;
 const { results, resultDigest } = buildUtf8ResultAndDigest('result');
 const EIP712DOMAIN_SEPARATOR = 'EIP712DOMAIN_SEPARATOR';
 const BALANCES = 'm_balances';
@@ -245,7 +244,7 @@ describe('IexecPocoBoostDelegate', function () {
             await expectOrderConsumed(iexecPocoBoostInstance, datasetOrderHash, undefined);
             await expectOrderConsumed(iexecPocoBoostInstance, workerpoolOrderHash, undefined);
             await expectOrderConsumed(iexecPocoBoostInstance, requestOrderHash, undefined);
-            await time.setNextBlockTimestamp(startTime);
+            const startTime = await setNextBlockTimestamp();
 
             await expect(
                 iexecPocoBoostInstance.matchOrdersBoost(
@@ -1201,7 +1200,7 @@ describe('IexecPocoBoostDelegate', function () {
             await signOrders(domain, orders, accounts);
             const dealId = getDealId(domain, requestOrder, taskIndex);
             const taskId = getTaskId(dealId, taskIndex);
-            await time.setNextBlockTimestamp(startTime);
+            const startTime = await setNextBlockTimestamp();
             await iexecPocoBoostInstance.matchOrdersBoost(
                 appOrder,
                 datasetOrder,
@@ -1384,7 +1383,7 @@ describe('IexecPocoBoostDelegate', function () {
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
                 buildCompatibleOrders(entriesAndRequester, dealTagTee);
             await signOrders(domain, orders, accounts);
-            await time.setNextBlockTimestamp(startTime);
+            const startTime = await setNextBlockTimestamp();
             await iexecPocoBoostInstance.matchOrdersBoost(
                 appOrder,
                 datasetOrder,
@@ -1588,7 +1587,7 @@ describe('IexecPocoBoostDelegate', function () {
             await expectFrozen(iexecPocoBoostInstance, requester.address, initialRequesterFrozen);
             const dealId = getDealId(domain, requestOrder, taskIndex);
             const taskId = getTaskId(dealId, taskIndex);
-            await time.setNextBlockTimestamp(startTime);
+            const startTime = await setNextBlockTimestamp();
             await iexecPocoBoostInstance.matchOrdersBoost(
                 appOrder,
                 datasetOrder,
@@ -1696,7 +1695,7 @@ describe('IexecPocoBoostDelegate', function () {
                 buildCompatibleOrders(entriesAndRequester, standardDealTag);
             await signOrders(domain, orders, accounts);
             const dealId = getDealId(domain, requestOrder, taskIndex);
-            await time.setNextBlockTimestamp(startTime);
+            const startTime = await setNextBlockTimestamp();
             await iexecPocoBoostInstance.matchOrdersBoost(
                 appOrder,
                 datasetOrder,
@@ -1715,6 +1714,19 @@ describe('IexecPocoBoostDelegate', function () {
         });
     });
 });
+
+/**
+ * Mine the next block with a timestamp corresponding to an arbitrary but known
+ * date in the future (10 seconds later).
+ * It fixes the `Timestamp is lower than the previous block's timestamp` error.
+ * e.g: This Error has been seen when running tests with `npm run coverage`.
+ * @returns timestamp of the next block.
+ */
+async function setNextBlockTimestamp() {
+    const startTime = (await time.latest()) + 10;
+    await time.setNextBlockTimestamp(startTime);
+    return startTime;
+}
 
 async function expectOrderConsumed(
     iexecPocoInstance: MockContract<IexecPocoBoostDelegate>,