+++ CHANGELOG.md
@@ -47,7 +47,7 @@
     - Handle callback. (#183)
 - Implement `claimBoost`:
     - Refund requester. (#198)
-- Refactor tests. (#177, #178, #179, #181, #188)
+- Refactor tests. (#177, #178, #179, #181, #188, #200)
 - Add mocha to use test explorer. (#180)
 
 ## v5.4.2